HBP-87. Move tests from tvb_nest at top-level under tests folder.
