+++ tvb_multiscale/tests/tvb_nest/test_simulator.py
